change value of counter to integer instead of smallInteger

+++ database/migrations/2020_09_17_071029_create_car_physics_table.php
@@ -17,7 +17,7 @@ class CreateCarPhysicsTable extends Migration
     {
         Schema::create('car_physics', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('counter');
+            $table->integer('counter');
             $table->timestamps();
         });
